+++ src/components/taskListCommon/task-list-bulk-actions-bar/task-list-bulk-actions-bar.tsx
@@ -52,7 +52,7 @@ import logger from '@/utils/errorLogger';
 import ConvertToSubtaskDrawer from '@/components/task-list-common/convert-to-subtask-drawer/convert-to-subtask-drawer';
 import { fetchLabels } from '@/features/taskAttributes/taskLabelSlice';
 import { useAuthService } from '@/hooks/useAuth';
-import CustomColumnModal from '@/pages/projects/projectView/taskList/taskListTable/custom-columns/custom-column-modal/custom-column-modal';
+import CustomColumnModal from '@/pages/projects/projectView/taskList/task-list-table/custom-columns/custom-column-modal/custom-column-modal';
 import { checkTaskDependencyStatus } from '@/utils/check-task-dependency-status';
 import alertService from '@/services/alerts/alertService';
