+++ src/pages/projects/projectView/taskList/groupTables/priorityTables/PriorityGroupTables.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TaskPriorityType, TaskType } from '../../../../../../types/task.types';
 import { Flex } from 'antd';
-import TaskListTableWrapper from '../../taskListTable/TaskListTableWrapper';
+import TaskListTableWrapper from '../../task-list-table/task-list-table-wrapper/task-list-table-wrapper';
 import { useAppSelector } from '../../../../../../hooks/useAppSelector';
 import { getPriorityColor } from '../../../../../../utils/getPriorityColors';
 import { IProjectTask } from '@/types/project/projectTasksViewModel.types';
