+++ src/pages/projects/projectView/taskList/task-list-table/task-list-table-cells/task-list-progress-cell/task-progress-cell.css
